+++ src/constants.lua
@@ -4,6 +4,10 @@ ADDON_NAME = "RaidNotesTBC"
 DEBUG_MODE = true
 
 -- set to true to enable test mode, must also enable debug_mode
-TEST_MODE = DEBUG_MODE and false
+TEST_MODE = DEBUG_MODE and true
 TEST_ZONE = "Sunwell Plateau"
-TEST_TARGET = "M'uru"
+TEST_TARGET = "Lady Sacrolash"
+TEST_INSTANCE_INFO = {
+    { "The Sunwell", 2 },
+    { "Black Temple", 3 }
+}